pkg/endpoint.EndpointChange: accept labels, container name and container ID
In order to modify an already running endpoint, cilium-docker plugin
will need to modify these fields from the endpoint structure. This is
because cilium-docker is the entity in charge to watch for docker
events and therefore will be able to know which labels are assigned for
each container started in a docker. Once it receive a new "start" event,
it will send a EndpointChange to cilium-agent so that cilium-agent is
aware of the labels associated with the recently endpoint created.

from endpoint request changes

Signed-off-by: André Martins <andre@cilium.io>